// src/components/ReferanslarSayfasiReferanslar/types.ts
export interface Props {
  sectionAnchorId?: string;
  reference11QuoteHtml?: string;
  reference6ImageUrl?: IkasImage | null;
  reference6DetailsHtml?: string;
  reference6NameText?: string;
  reference6QuoteHtml?: string;
  reference6TypeText?: string;
  reference11ImageUrl?: IkasImage | null;
  reference7TypeText?: string;
  reference5ImageUrl?: IkasImage | null;
  reference5NameText?: string;
  reference5QuoteHtml?: string;
  reference5TypeText?: string;
  reference4ImageUrl?: IkasImage | null;
  reference4DetailsHtml?: string;
  reference4NameText?: string;
  reference4QuoteHtml?: string;
  reference5DetailsHtml?: string;
  reference4TypeText?: string;
  reference7QuoteHtml?: string;
  reference7DetailsHtml?: string;
  reference10DetailsHtml?: string;
  reference10NameText?: string;
  reference10QuoteHtml?: string;
  reference10TypeText?: string;
  reference9ImageUrl?: IkasImage | null;
  reference9DetailsHtml?: string;
  reference9NameText?: string;
  reference7NameText?: string;
  reference9QuoteHtml?: string;
  reference11TypeText?: string;
  reference8ImageUrl?: IkasImage | null;
  reference8DetailsHtml?: string;
  reference8NameText?: string;
  reference8QuoteHtml?: string;
  reference8TypeText?: string;
  reference7ImageUrl?: IkasImage | null;
  reference9TypeText?: string;
  reference11NameText?: string;
  reference10ImageUrl?: IkasImage | null;
  reference2ImageUrl?: IkasImage | null;
  reference1ImageUrl?: IkasImage | null;
  reference11DetailsHtml?: string;
  reference2TypeText?: string;
  reference2QuoteHtml?: string;
  reference1NameText?: string;
  reference2NameText?: string;
  reference1QuoteHtml?: string;
  reference2DetailsHtml?: string;
  reference3ImageUrl?: IkasImage | null;
  reference1TypeText?: string;
  reference3TypeText?: string;
  reference1DetailsHtml?: string;
  reference3DetailsHtml?: string;
  referencesAriaLabel?: string;
  reference3QuoteHtml?: string;
  reference3NameText?: string;
  reference9TypeTextEn?: string;
  reference7DetailsHtmlEn?: string;
  reference10NameTextEn?: string;
  reference10QuoteHtmlEn?: string;
  reference1NameTextEn?: string;
  reference11TypeTextEn?: string;
  reference10TypeTextEn?: string;
  reference8TypeTextEn?: string;
  reference1QuoteHtmlEn?: string;
  reference9DetailsHtmlEn?: string;
  reference8QuoteHtmlEn?: string;
  reference9NameTextEn?: string;
  reference9QuoteHtmlEn?: string;
  reference8NameTextEn?: string;
  reference1TypeTextEn?: string;
  reference8DetailsHtmlEn?: string;
  reference11DetailsHtmlEn?: string;
  referencesAriaLabelEn?: string;
  reference3DetailsHtmlEn?: string;
  reference7QuoteHtmlEn?: string;
  reference1DetailsHtmlEn?: string;
  reference4TypeTextEn?: string;
  reference3NameTextEn?: string;
  reference4QuoteHtmlEn?: string;
  reference4NameTextEn?: string;
  reference3QuoteHtmlEn?: string;
  reference4DetailsHtmlEn?: string;
  reference7NameTextEn?: string;
  reference3TypeTextEn?: string;
  reference5TypeTextEn?: string;
  reference10DetailsHtmlEn?: string;
  reference11NameTextEn?: string;
  reference11QuoteHtmlEn?: string;
  reference5DetailsHtmlEn?: string;
  reference7TypeTextEn?: string;
  reference2TypeTextEn?: string;
  reference6DetailsHtmlEn?: string;
  reference5NameTextEn?: string;
  reference6NameTextEn?: string;
  reference5QuoteHtmlEn?: string;
  reference6QuoteHtmlEn?: string;
  reference6TypeTextEn?: string;
  reference2NameTextEn?: string;
  reference2DetailsHtmlEn?: string;
  reference2QuoteHtmlEn?: string;
  paddingTop?: number;
  paddingBottom?: number;
  lineColor?: string;
  mutedTextColor?: string;
  textColor?: string;
  backgroundColor?: string;
  maxWidth?: number;
  showReference10?: boolean;
  showReference5?: boolean;
  showReference8?: boolean;
  showReference7?: boolean;
  showReference6?: boolean;
  showReference4?: boolean;
  showReference3?: boolean;
  showReference2?: boolean;
  showReference1?: boolean;
  showReferenceWall?: boolean;
  showReference9?: boolean;
  showReference11?: boolean;
}